+++ CHANGELOG.md
@@ -10,9 +10,11 @@ All notable changes to this project will be documented in this file.
 
 ### Improvements
 - 'sendInternalRequest' returns now the HTTP response
+- FlowConfig handling
 - Minor docu improvements
 
 ### Bugfix
+- FlowConfig priority was not instance specific
 - Legacy bindings of ValueDisplay elements and FileUpload feature within UI did not work if deployed with VS Code AppSpace SDK
 - UI differs if deployed via Appstudio or VS Code AppSpace SDK
 - Fullscreen icon of iFrame was visible